Fix singular `ntPrincipalName`
An erratum was published mentioning that an `ntPrincipalName` has
to be singular and doesn't support an array. The full text:

The documentation indicated that all the keys in the SubjectAltName
value could be either string or array types. The ntPrincipalName
cannot be an array and must be a string. This has been clarified
in the description. Note that the type field for the rfc822Name,
dNSName, and uniformResourceIdentifier still indicates these are
strings. This has not been corrected as the schema does not support
polymorphic types.